requirements.txt
streamlit==0.84.1
numpy==1.20.0
Pillow
